Modularize view options, unify patterns for violin and scatter plots
